requirements.txt
Django==1.2.1
mutagen==1.19
# gdal==1.7.1
south==0.7.1
django-guardian==0.1.0
django-tagging==0.3.1
django-imagekit==0.3.3
django-voting==0.1
django-reversion==1.3.1
django-annoying==0.7.4
django-reviews==0.2.1
django-debug-toolbar==0.8.3
django-extensions==0.5
django-profiles==0.2
django-ajax-selects==1.1.3
django-request==0.23
vobject==0.8.1c
python-dateutil==1.5

-e git+http://github.com/ericflo/django-threadedcomments.git@ded46fc5bfd9a9d37a15351aeafec58df4ae9301#egg=django-threadedcomments
-e hg+http://bitbucket.org/ubernostrum/django-registration#egg=django-registration
-e git+http://github.com/mintchaos/django_compressor.git#egg=django-compressor
-e git+http://github.com/pydanny/django-uni-form.git#egg=django-uni-form
-e git+http://github.com/flashingpumpkin/django-socialregistration.git@v0.4-tag#egg=django-socialregistration
-e git+http://github.com/glibersat/django-agenda.git#egg=django-agenda
-e git+http://github.com/justquick/django-activity-stream.git@bf57186e09605ece25d2#egg=django-activity-stream
-e git+http://github.com/liberation/django-notification.git@38ad9cd9beba457e6e09fc619dd278b448781768#egg=django-notification
-e git+http://github.com/kmmbvnr/django-fsm.git#egg=django-fsm
